# api/openapi/contract.base.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "CPRa SDK draft management contract",
    "version": "0.1.0-rc.1",
    "description": "Draft contract. V2 server qualification is pending. Existing v1 is separate."
  },
  "paths": {
...
    "/api/v2/operations": {
      "post": {
        "operationId": "CreateOperation",
        "summary": "CreateOperation",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./models.json#/components/schemas/Operation"
                }
              }
            }
          },
          "default": {
            "description": "Problem details",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "./models.json#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "./models.json#/components/schemas/OperationCreateRequest"
              }
            }
          }
        }
      },
      "get": {
        "operationId": "ListOperations",
        "summary": "ListOperations",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./models.json#/components/schemas/OperationList"
                }
              }
            }
          },
          "default": {
            "description": "Problem details",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "./models.json#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "selector",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "monitorID",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v2/operations/{id}/validate": {
      "post": {
        "operationId": "ValidateOperation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Original validation request accepted or reconciled. Poll GET validation; do not resubmit to poll.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./models.json#/components/schemas/Operation"
                }
              }
            }
          },
          "default": {
            "description": "Problem details",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "./models.json#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/operations/{id}/validation": {
      "get": {
        "operationId": "GetOperationValidation",
        "summary": "Read retained validation results",
        "description": "Reads one immutable sealed verdict and bounded original-order result page. 409 validationPending includes Retry-After; interrupted/canceled/expired validation without a verdict is distinct. This read never requests compilation or mutation.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 4096
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 500,
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sealed validation result page",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "./models.json#/components/schemas/ValidationResultPage"
                }
              }
            }
          },
          "default": {
            "description": "Problem details",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "./models.json#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  }
}